pipe_plugins2: Rename files after morning talks.
