+++ CMakeLists.txt
@@ -1408,8 +1408,8 @@ set(INTERNAL_COLLECTORS_FILES
 )
 
 set(INTERCOMMUNICATION_COLLECTORS_FILES
-        src/collectors/collectors-ipc/collectors-ipc.c
-        src/collectors/collectors-ipc/collectors-ipc.h
+        src/collectors/collectors-ipc/ebpf-ipc.c
+        src/collectors/collectors-ipc/ebpf-ipc.h
 )
 
 set(PLUGINSD_PLUGIN_FILES
